[Pull]

+++ HMES.Data/DTO/RequestModel/CartReqModel.cs
@@ -9,12 +9,12 @@ public class CartItemCreateDto
 {
     public Guid ProductId { get; set; }
     public decimal UnitPrice { get; set; }
-    public decimal Quantity { get; set; }
+    public int Quantity { get; set; }
 }
 
 public class CartItemUpdateDto
 {
     public Guid ProductId { get; set; }
     public decimal UnitPrice { get; set; }
-    public decimal Quantity { get; set; }
+    public int Quantity { get; set; }
 }
